toolkit/spatial-navigation/tests/chrome/test_snav_textFields.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css type="text/css"?>

<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=436084
-->

<window title="Mozilla Bug 288254"
  xmlns:html="http://www.w3.org/1999/xhtml"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
  onload="onLoad();">

  <script type="application/javascript" src="chrome://mochikit/content/MochiKit/packed.js"></script>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>      
  <script type="text/javascript" src="SpatialNavUtils.js"></script>

<body id="some-content" xmlns="http://www.w3.org/1999/xhtml">

<p>
<a id="start" href="https://bugzilla.mozilla.org/show_bug.cgi?id=436084">Mozilla Bug 436084</a>
</p>

<p>
<input id="textinput" value="abcdefghijklmnopqrstuvwxyz" type="text"></input>
</p>

<p>
<textarea id="textarea" name="textarea" cols="30" rows="3" wrap="hard">The Book of Mozilla</textarea>
</p>

<p>
<a id="end" href="https://bugzilla.mozilla.org/show_bug.cgi?id=436084">Mozilla Bug 436084</a>
</p>

</body>


<script class="testbody" type="application/javascript">
<![CDATA[
Components.utils.import("resource://gre/modules/SpatialNavigation.js");

var moveTable = [
["DOWN", "textinput"],
["DOWN", "textinput"],
["DOWN", "textarea"],
["DOWN", "textarea"],
["DOWN", "end"],
["UP", "textarea"],
["UP", "textarea"],
["UP", "textinput"],
["UP", "textinput"],
["UP", "start"],
["DOWN", "textinput"],
["LEFT", "textinput"],


["RIGHT", "textinput"],
["RIGHT", "textinput"],
["RIGHT", "textinput"],
["RIGHT", "textinput"],
["RIGHT", "textinput"],
["RIGHT", "textinput"],
["RIGHT", "textinput"],
["RIGHT", "textinput"],
["RIGHT", "textinput"],
["RIGHT", "textinput"],
["RIGHT", "textinput"],
["RIGHT", "textinput"],
["RIGHT", "textinput"],
["RIGHT", "textinput"],
["RIGHT", "textinput"],
["RIGHT", "textinput"],
["RIGHT", "textinput"],
["RIGHT", "textinput"],
["RIGHT", "textinput"],
["RIGHT", "textinput"],
["RIGHT", "textinput"],
["RIGHT", "textinput"],
["RIGHT", "textinput"],
["RIGHT", "textinput"],
["RIGHT", "textinput"],
["RIGHT", "textinput"],
["RIGHT", "textarea"],
["LEFT", "textarea"],
["RIGHT", "textarea"],
["RIGHT", "textarea"],
["RIGHT", "textarea"],
["RIGHT", "textarea"],
["RIGHT", "textarea"],
["RIGHT", "textarea"],
["RIGHT", "textarea"],
["RIGHT", "textarea"],
["RIGHT", "textarea"],
["RIGHT", "textarea"],
["RIGHT", "textarea"],
["RIGHT", "textarea"],
["RIGHT", "textarea"],
["RIGHT", "textarea"],
["RIGHT", "textarea"],
["RIGHT", "textarea"],
["RIGHT", "textarea"],
["RIGHT", "textarea"],
["RIGHT", "textarea"],
["RIGHT", "end"],
["UP", "textarea"],
["DONE", "DONE"],
];

var prefs = [
["enabled",           "bool", true],
["xulContentEnabled", "bool", true],
["keyCode.modifier",  "char", "none"],
];

function onLoad()
{
    prepareTest(prefs);

    // starting the test itself.
    var x = document.getElementById("some-content");
    SpatialNavigation.init(x);
    
    // get to a known place.
    document.getElementById("start").focus();

    testMoves(moveTable);

    SimpleTest.waitForExplicitFinish();

    SpatialNavigation.uninit();
}

]]></script>
</window>
